# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Task Assistant Worker API
  description: |
    Cloudflare Workers + Hono 后端 HTTP API（PRD v1.2 / 技术方案 v1.4 对齐）。

    **鉴权**：多数 `/api/*` 需 `Authorization: Bearer <token>`。`token` 为 **JWT**（`POST /api/auth/login` 签发，`sub` = `users.id`）或开发用 **明文用户 id**（如 `local-dev-user`，须存在于 D1）。

    **SSE 对话**：`POST /api/chat/stream` 返回 `text/event-stream`；帧格式为 `event: <名称>` + `data: <JSON>` + 空行。事件类型见下方该路径说明及 `GET /health/sse-chat`。

    **管理 Prompt**：`/api/prompts/*` 需配置 Worker 环境变量 `ADMIN_API_SECRET`，并使用 `Authorization: Bearer <ADMIN_API_SECRET>` 或 `X-Admin-Token: <ADMIN_API_SECRET>`。
  version: 1.2.0
servers:
  - url: http://127.0.0.1:8787
    description: 本地 wrangler dev
  - url: https://{workerSubdomain}.workers.dev
    description: Cloudflare Workers 默认子域（部署后替换）
    variables:
      workerSubdomain:
        default: your-worker
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT（login）或明文 users.id（仅开发）
  schemas:
...
    AuthUser:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, nullable: true }
        ai_nickname: { type: string }
    LoginRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: 显示名；已存在则登录，否则注册
        email:
          type: string
          nullable: true
          description: 可选；新用户可填，须全局唯一
    LoginResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/AuthUser' }
        is_new_user: { type: boolean }
    ProfileExistsResponse:
      type: object
      properties:
        exists: { type: boolean }
...
paths:
...
  /api/auth/profile-exists:
    get:
      tags: [Auth]
      summary: 查询昵称是否已被占用（不签发 token）
      parameters:
        - name: name
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfileExistsResponse' }
        '400':
          description: 缺少 name
  /api/auth/login:
    post:
      tags: [Auth]
      summary: 注册或登录（JWT）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '409':
          description: 邮箱冲突（新用户）
        '503':
          description: 未配置 JWT_SECRET
...
  /api/files/upload:
    post:
      tags: [Files]
      security: [{ bearerAuth: [] }]
      summary: 小文件 multipart/form 上传
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                semantic_type: { type: string }
                folder_path: { type: string }
                tags:
                  type: string
                  description: JSON 数组字符串
      responses:
        '201':
          description: 已创建
  /api/files/initiate-multipart:
    post:
      tags: [Files]
      security: [{ bearerAuth: [] }]
      summary: 大文件分片上传 — 初始化
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [filename, size]
              properties:
                filename: { type: string }
                original_name: { type: string }
                mime_type: { type: string }
                size: { type: number }
                semantic_type: { type: string, nullable: true }
                folder_path: { type: string }
                tags:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: upload_id、r2_key、part_urls
  /api/files/complete-multipart:
    post:
      tags: [Files]
      security: [{ bearerAuth: [] }]
      summary: 大文件分片上传 — 完成
      responses:
        '200':
          description: OK
  /api/files/{id}:
    get:
      tags: [Files]
      security: [{ bearerAuth: [] }]
      summary: 单文件元数据（含异步处理状态 processed）
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '404':
          description: 不存在
    delete:
      tags: [Files]
      security: [{ bearerAuth: [] }]
      summary: 删除文件
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
  /api/files/{id}/rename:
    put:
      tags: [Files]
      security: [{ bearerAuth: [] }]
      summary: 重命名
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [new_name]
              properties:
                new_name: { type: string }
      responses:
        '200':
          description: OK
  /api/files/{id}/semantic-type:
    put:
      tags: [Files]
      security: [{ bearerAuth: [] }]
      summary: 更新语义类型
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                semantic_type: { type: string, nullable: true }
      responses:
        '200':
          description: OK
  /api/files/{id}/tags:
    put:
      tags: [Files]
      security: [{ bearerAuth: [] }]
      summary: 更新标签
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [tags]
              properties:
                tags:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: OK
  /api/files/{id}/download:
    get:
      tags: [Files]
      security: [{ bearerAuth: [] }]
      summary: 获取预签名下载 URL
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string }
  /api/files/{id}/retry-process:
    post:
      tags: [Files]
      security: [{ bearerAuth: [] }]
      summary: 异步解析/向量化失败后重试
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
...